requirements.txt
# MLOps Image Classification - Requirements Principales
# Pour l'entraînement du modèle et l'exploration des données

# Machine Learning & Deep Learning
torch==2.1.0
torchvision==0.16.0
fastai==2.7.12
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# Computer Vision
Pillow==10.0.1
opencv-python==4.8.1.78

# MLflow & Tracking
mlflow==2.8.1
boto3==1.34.0  # Pour S3/Minio

# Visualisation
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Data Processing
tqdm==4.66.1
pathlib2==2.3.7

# Database
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23

# Utilities
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Load Testing
locust==2.17.0

# Monitoring
prometheus-client==0.19.0
psutil==5.9.6

# Jupyter (optionnel pour l'exploration)
jupyter==1.0.0
ipykernel==6.26.0